+++ include/roq/fbs/api.hpp
@@ -205,6 +205,59 @@ inline const char *EnumNameBufferCapacity(BufferCapacity e) {
   return EnumNamesBufferCapacity()[index];
 }
 
+enum class CancelAllOrdersFilter : uint32_t {
+  Undefined = 0,
+  Account = 1,
+  Exchange = 2,
+  Symbol = 4,
+  Strategy = 8,
+  Side = 16,
+  MIN = Undefined,
+  MAX = Side
+};
+
+inline const CancelAllOrdersFilter (&EnumValuesCancelAllOrdersFilter())[6] {
+  static const CancelAllOrdersFilter values[] = {
+    CancelAllOrdersFilter::Undefined,
+    CancelAllOrdersFilter::Account,
+    CancelAllOrdersFilter::Exchange,
+    CancelAllOrdersFilter::Symbol,
+    CancelAllOrdersFilter::Strategy,
+    CancelAllOrdersFilter::Side
+  };
+  return values;
+}
+
+inline const char * const *EnumNamesCancelAllOrdersFilter() {
+  static const char * const names[18] = {
+    "Undefined",
+    "Account",
+    "Exchange",
+    "",
+    "Symbol",
+    "",
+    "",
+    "",
+    "Strategy",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "Side",
+    nullptr
+  };
+  return names;
+}
+
+inline const char *EnumNameCancelAllOrdersFilter(CancelAllOrdersFilter e) {
+  if (::flatbuffers::IsOutRange(e, CancelAllOrdersFilter::Undefined, CancelAllOrdersFilter::Side)) return "";
+  const size_t index = static_cast<size_t>(e);
+  return EnumNamesCancelAllOrdersFilter()[index];
+}
+
 enum class Category : uint8_t {
   Undefined = 0,
   Public = 1,
@@ -4386,7 +4439,8 @@ struct GatewaySettings FLATBUFFERS_FINAL_CLASS : private ::flatbuffers::Table {
     VT_MBP_TICK_SIZE_MULTIPLIER = 18,
     VT_MBP_MIN_TRADE_VOL_MULTIPLIER = 20,
     VT_OMS_REQUEST_ID_TYPE = 22,
-    VT_MBP_CHECKSUM = 24
+    VT_MBP_CHECKSUM = 24,
+    VT_OMS_CANCEL_ALL_ORDERS = 26
   };
   uint64_t supports() const {
     return GetField<uint64_t>(VT_SUPPORTS, 0);
@@ -4418,6 +4472,9 @@ struct GatewaySettings FLATBUFFERS_FINAL_CLASS : private ::flatbuffers::Table {
   bool mbp_checksum() const {
     return GetField<uint8_t>(VT_MBP_CHECKSUM, 0) != 0;
   }
+  uint64_t oms_cancel_all_orders() const {
+    return GetField<uint64_t>(VT_OMS_CANCEL_ALL_ORDERS, 0);
+  }
   bool Verify(::flatbuffers::Verifier &verifier) const {
     return VerifyTableStart(verifier) &&
            VerifyField<uint64_t>(verifier, VT_SUPPORTS, 8) &&
@@ -4430,6 +4487,7 @@ struct GatewaySettings FLATBUFFERS_FINAL_CLASS : private ::flatbuffers::Table {
            VerifyField<double>(verifier, VT_MBP_MIN_TRADE_VOL_MULTIPLIER, 8) &&
            VerifyField<uint8_t>(verifier, VT_OMS_REQUEST_ID_TYPE, 1) &&
            VerifyField<uint8_t>(verifier, VT_MBP_CHECKSUM, 1) &&
+           VerifyField<uint64_t>(verifier, VT_OMS_CANCEL_ALL_ORDERS, 8) &&
            verifier.EndTable();
   }
 };
@@ -4468,6 +4526,9 @@ struct GatewaySettingsBuilder {
   void add_mbp_checksum(bool mbp_checksum) {
     fbb_.AddElement<uint8_t>(GatewaySettings::VT_MBP_CHECKSUM, static_cast<uint8_t>(mbp_checksum), 0);
   }
+  void add_oms_cancel_all_orders(uint64_t oms_cancel_all_orders) {
+    fbb_.AddElement<uint64_t>(GatewaySettings::VT_OMS_CANCEL_ALL_ORDERS, oms_cancel_all_orders, 0);
+  }
   explicit GatewaySettingsBuilder(::flatbuffers::FlatBufferBuilder &_fbb)
         : fbb_(_fbb) {
     start_ = fbb_.StartTable();
@@ -4490,8 +4551,10 @@ inline ::flatbuffers::Offset<GatewaySettings> CreateGatewaySettings(
     double mbp_tick_size_multiplier = std::numeric_limits<double>::quiet_NaN(),
     double mbp_min_trade_vol_multiplier = std::numeric_limits<double>::quiet_NaN(),
     roq::fbs::RequestIdType oms_request_id_type = roq::fbs::RequestIdType::Undefined,
-    bool mbp_checksum = false) {
+    bool mbp_checksum = false,
+    uint64_t oms_cancel_all_orders = 0) {
   GatewaySettingsBuilder builder_(_fbb);
+  builder_.add_oms_cancel_all_orders(oms_cancel_all_orders);
   builder_.add_mbp_min_trade_vol_multiplier(mbp_min_trade_vol_multiplier);
   builder_.add_mbp_tick_size_multiplier(mbp_tick_size_multiplier);
   builder_.add_supports(supports);