+++ include/roq/gateway_settings.hpp
@@ -9,6 +9,7 @@
 #include <fmt/compile.h>
 #include <fmt/format.h>
 
+#include "roq/cancel_all_orders_filter.hpp"
 #include "roq/event.hpp"
 #include "roq/mask.hpp"
 #include "roq/name.hpp"
@@ -21,16 +22,17 @@ namespace roq {
 
 //! Gateway settings
 struct ROQ_PUBLIC GatewaySettings final {
-  Mask<SupportType> supports;                  //!< Supported update types
-  uint16_t mbp_max_depth = {};                 //!< MBP max depth
-  double mbp_tick_size_multiplier = NaN;       //!< MBP multiplier used to manage prices as integer
-  double mbp_min_trade_vol_multiplier = NaN;   //!< MBP multiplier used to manage quantities as integer
-  bool mbp_allow_remove_non_existing = false;  //!< MBP allow remove operation on non-existing level?
-  bool mbp_allow_price_inversion = false;      //!< MBP allow price inversion?
-  bool mbp_checksum = false;                   //!< MBP compute checksum?
-  bool oms_download_has_state = false;         //!< OMS download includes state information?
-  bool oms_download_has_routing_id = false;    //!< OMS download includes routing_id?
-  RequestIdType oms_request_id_type = {};      //!< OMS request identifier type
+  Mask<SupportType> supports;                         //!< Supported update types
+  uint16_t mbp_max_depth = {};                        //!< MBP max depth
+  double mbp_tick_size_multiplier = NaN;              //!< MBP multiplier used to manage prices as integer
+  double mbp_min_trade_vol_multiplier = NaN;          //!< MBP multiplier used to manage quantities as integer
+  bool mbp_allow_remove_non_existing = false;         //!< MBP allow remove operation on non-existing level?
+  bool mbp_allow_price_inversion = false;             //!< MBP allow price inversion?
+  bool mbp_checksum = false;                          //!< MBP compute checksum?
+  bool oms_download_has_state = false;                //!< OMS download includes state information?
+  bool oms_download_has_routing_id = false;           //!< OMS download includes routing_id?
+  RequestIdType oms_request_id_type = {};             //!< OMS request identifier type
+  Mask<CancelAllOrdersFilter> oms_cancel_all_orders;  //!< Supported filter types when using CancelAllOrders
 };
 
 template <>
@@ -63,7 +65,8 @@ struct fmt::formatter<roq::GatewaySettings> {
         R"(mbp_checksum={}, )"
         R"(oms_download_has_state={}, )"
         R"(oms_download_has_routing_id={}, )"
-        R"(oms_request_id_type={})"
+        R"(oms_request_id_type={}, )"
+        R"(oms_cancel_all_orders={})"
         R"(}})"_cf,
         value.supports,
         value.mbp_max_depth,
@@ -74,7 +77,8 @@ struct fmt::formatter<roq::GatewaySettings> {
         value.mbp_checksum,
         value.oms_download_has_state,
         value.oms_download_has_routing_id,
-        value.oms_request_id_type);
+        value.oms_request_id_type,
+        value.oms_cancel_all_orders);
   }
 };